Renaming group to “Configuration”
and adding header comments to all Swift files.
