+++ Swift/Sources/Configuration/Prayer.swift
@@ -1,5 +1,5 @@
 //
-//  TimeComponents.swift
+//  Prayer.swift
 //  Adhan
 //
 //  Copyright © 2018 Batoul Apps. All rights reserved.
@@ -25,20 +25,12 @@
 
 import Foundation
 
-struct TimeComponents {
-    let hours: Int
-    let minutes: Int
-    let seconds: Int
-
-    func dateComponents(_ date: DateComponents) -> DateComponents {
-        var comps = DateComponents()
-        comps.year = date.year
-        comps.month = date.month
-        comps.day = date.day
-        comps.hour = self.hours
-        comps.minute = self.minutes
-        comps.second = self.seconds
-
-        return comps
-    }
+public enum Prayer {
+    case fajr
+    case sunrise
+    case dhuhr
+    case asr
+    case maghrib
+    case isha
+    case none
 }
